backend/requirement.txt
fastapi == 0.115.12
uvicorn[standard]
sqlmodel
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
anthropic